# apps/backend/documentation/todoManager.openapi.yml
openapi: 3.0.3
info:
  title: Todo Manager API
  description: |
    API responsible for managing production Demand planning for **Latinhas LLC**.

    Each Demand has a planning window (start and end dates), a status, and a list of items.
    Every item is associated with a SKU and stores the planned total in metric tons.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /demands/{demandId}:
    parameters:
      - $ref: "#/components/parameters/DemandId"
    get:
      tags:
        - Demands
      summary: Get a demand by ID
      description: Returns the full details of a specific demand.
      responses:
        "200":
          description: Demand retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags:
        - Demands
      summary: Update an existing demand
      description: Updates an existing demand with the provided data. Items are not modified in this operation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemandUpdatePayload"
      responses:
        "200":
          description: Demand updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags:
        - Demands
      summary: Delete a demand
      description: Deletes a demand identified by the provided ID.
      responses:
        "204":
          description: Demand deleted successfully.
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    DemandId:
      name: demandId
      in: path
      required: true
      description: Unique identifier of the demand.
      schema:
        type: string
        format: uuid
...
  schemas:
    DemandStatus:
      type: string
      description: Current status of the demand lifecycle.
      enum:
        - planning
        - in_progress
        - completed
    DemandItemInput:
      type: object
      required:
        - description
        - plannedTotalTons
      properties:
        description:
          type: string
          minLength: 1
          description: Short description of the planned item. Cannot be empty.
        plannedTotalTons:
          type: number
          format: float
          minimum: 0
          description: Planned production total for the item, measured in metric tons.
    DemandItem:
      allOf:
        - type: object
          required:
            - sku
          properties:
            sku:
              type: integer
              format: int32
              minimum: 1
              description: Unique SKU identifier (primary key) generated by the persistence layer.
        - $ref: "#/components/schemas/DemandItemInput"
        - type: object
          properties:
            producedTotalTons:
              type: number
              format: float
              minimum: 0
              description: Actual produced total for the item, when available.
...
    DemandSummary:
      type: object
      required:
        - id
        - title
        - status
        - startDate
        - endDate
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the demand.
        title:
          type: string
          description: Short name used to reference the demand.
          example: Q1/2025 Demand
        status:
          $ref: "#/components/schemas/DemandStatus"
        startDate:
          type: string
          format: date
          description: Planned start date using the ISO 8601 format (YYYY-MM-DD).
        endDate:
          type: string
          format: date
          description: Planned end date using the ISO 8601 format (YYYY-MM-DD).
    Demand:
      allOf:
        - $ref: "#/components/schemas/DemandSummary"
        - type: object
          required:
            - items
          properties:
            description:
              type: string
              description: Detailed description of the demand.
            items:
              type: array
              description: |
                Planned items for this demand. Each item includes its SKU and the planned total in metric tons.
              minItems: 1
              items:
                $ref: "#/components/schemas/DemandItem"
            createdAt:
              type: string
              format: date-time
              description: Timestamp indicating when the demand was created.
            updatedAt:
              type: string
              format: date-time
              description: Timestamp indicating the last update.
...
    DemandUpdatePayload:
      type: object
      required:
        - title
        - description
        - status
        - endDate
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 140
          description: Human-friendly title used to identify the demand.
        description:
          type: string
          minLength: 1
          description: Additional notes or business context for the demand. Cannot be empty.
        status:
          $ref: "#/components/schemas/DemandStatus"
        endDate:
          type: string
          format: date
          description: Planned end date using the ISO 8601 format (YYYY-MM-DD). Must be greater than or equal to the start date.
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        code:
          type: string
          description: Internal error code used to support troubleshooting.
          example: demand_not_found
        message:
          type: string
          description: Short human-readable description of the error.
        details:
          type: object
          additionalProperties: true
          description: Optional metadata that helps diagnose the issue.
    ValidationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            details:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
              description: Map of validation issues grouped by field.
  responses:
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Invalid input supplied for the requested operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
    InternalError:
      description: Unexpected error while processing the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"